backend/products/requirements.txt
flask
psycopg2-binary
flask-cors
prometheus-flask-exporter